change login function teacher and school have diffrent diffrent fuction i merg it in one only login function
