+++ controller/teacherController.js
@@ -30,38 +30,7 @@ const teacherRegister = async (req, res) => {
     res.status(500).json({ status: false, msg: "error to register" });
   }
 };
-const teacherLogin = async (req, res) => {
-  try {
-    const { email, password } = req.body;
-    const userDetails = await User.findOne({ email });
-
-    if (!userDetails) {
-      return res.status(400).json({ status: 400, msg: "User does not exist" });
-    }
-
-    const isMatch = await bcrypt.compare(password, userDetails.password);
-    if (!isMatch) {
-      return res.status(400).json({ status: 400, msg: "Invalid credentials" });
-    }
-
-    const user = {
-      id: userDetails.id,
-      name: userDetails.name,
-      email: userDetails.email,
-      role: userDetails.role,
-    };
-    const token = generateJWTToken(user);
-
-    return res
-      .status(200)
-      .json({ status: 200, msg: "User login successfully", user, token });
-  } catch (error) {
-    console.error("Login Error:", error);
-    res.status(500).json({ status: 500, msg: "Login Failed" });
-  }
-};
 
 module.exports = {
   teacherRegister,
-  teacherLogin,
 };
