+++ routes/teacherRoute.js
@@ -1,11 +1,6 @@
 const express = require("express");
-const { verifyToken } = require("../middleware/auth");
-const {
-  teacherRegister,
-  teacherLogin,
-} = require("../controller/teacherController");
+const { teacherRegister } = require("../controller/teacherController");
 const router = express.Router();
 
 router.post("/teacherRegister", teacherRegister);
-router.post("/teacherLogin", teacherLogin);
 module.exports = router;
