+++ server.js
@@ -2,6 +2,7 @@ const express = require("express");
 const dotenv = require("dotenv");
 const cors = require("cors");
 const connectDB = require("./config/db");
+const loginRoute = require("./routes/loginRoute");
 const schoolRoute = require("./routes/schoolRoute");
 const teacherRoute = require("./routes/teacherRoute");
 dotenv.config();
@@ -9,6 +10,7 @@ connectDB();
 const app = express();
 app.use(express.json());
 app.use(cors());
+app.use("/api/login", loginRoute);
 app.use("/api/school", schoolRoute);
 app.use("/api/teacher", teacherRoute);
 app.listen(process.env.PORT, "0.0.0.0", () => {
